+++ examples/data_disk/providers.tf
@@ -1,11 +1,3 @@
-terraform {
-  required_providers {
-    azurerm = {
-      source  = "hashicorp/azurerm"
-      version = ">=3.1.0"
-    }
-  }
-
-  required_version = ">= 1.0.0"
-
+provider "azurerm" {
+  features {}
 }