+++ examples/identity_user_assigned/main.tf
@@ -1,6 +1,7 @@
 resource "azurerm_resource_group" "vmss" {
   name     = var.vmss_resource_group_name
   location = var.vmss_location
+  tags     = var.tags
 }
 
 resource "azurerm_virtual_network" "vmss" {
@@ -18,19 +19,25 @@ resource "azurerm_subnet" "agents" {
   virtual_network_name = azurerm_virtual_network.vmss.name
 }
 
+resource "azurerm_user_assigned_identity" "vmss" {
+  name                = var.user_assigned_identity_name
+  resource_group_name = azurerm_resource_group.vmss.name
+  location            = azurerm_resource_group.vmss.location
+  tags                = var.tags
+}
+
 module "vmss" {
   source = "../.."
 
-  vmss_os                   = var.vmss_os
-  vmss_name                 = var.vmss_name
-  vmss_computer_name_prefix = var.vmss_computer_name_prefix
-  vmss_resource_group_name  = azurerm_resource_group.vmss.name
-  vmss_subnet_id            = azurerm_subnet.agents.id
-  vmss_admin_password       = var.vmss_admin_password
-  vmss_se_enabled           = var.vmss_se_enabled
-  # passing user_data that contains a JSON configuration for installs
-  vmss_user_data              = filebase64("${path.module}/user_data.json")
-  vmss_win_se_settings_script = var.vmss_win_se_settings_script
-  vmss_win_se_settings_data   = var.vmss_win_se_settings_data
-  vmss_win_se_settings        = var.vmss_win_se_settings
+  vmss_name                = var.vmss_name
+  vmss_resource_group_name = azurerm_resource_group.vmss.name
+  vmss_subnet_id           = azurerm_subnet.agents.id
+  vmss_admin_password      = var.vmss_admin_password
+
+  vmss_identity = {
+    type         = "UserAssigned"
+    identity_ids = [azurerm_user_assigned_identity.vmss.id]
+  }
+
+  tags = var.tags
 }