+++ versions.tf
@@ -2,9 +2,9 @@ terraform {
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = ">=3.1.0, <4.0.0"
+      version = "~> 4.0"
     }
   }
 
-  required_version = ">= 1.0.0"
+  required_version = ">= 1.10.0"
 }